Add survival_analyses.R
This script will take "model_inputs.rds" and output:

1. Table 1
2. Table with univariate R vs L and 5-yr survival Cox PH results
3. Schoenfeld residual plots for the above analysis
4. Table with multivariate R vs L and 5-yr survival Cox PH results
5. Schoenfeld residual plots for the above analysis
6. Table with lasso regression for 5-yr survival results
7. Plot of KM curve of R vs L and 5-yr survival

All tables are HTML tables. To use in Microsoft Word, it is best to open
them in a browser then copy and paste them into Word.

All plots are in png. Graphic size and format is easy to tweak by
changing the script.

All file outputs are put into "../output/" directory with self-explanatory
names.